登录态,以及登录后再请求 封装

+++ .eslintrc.js
@@ -26,7 +26,7 @@ module.exports = {
     // allow debugger during development
     'no-debugger': process.env.NODE_ENV === 'production' ? 2 : 0,
     "camelcase": [0, { "properties": "never" }],
-    "quotes": [0, "single"]
+    "quotes": [0, "double"]
   },
   globals: {
     App: true,
